+++ CHANGELOG.md
@@ -8,6 +8,9 @@ app.
 
 ## [Unreleased]
 
+- Made it possible to change tabs in the popup by swiping horizontally
+  thanks to [@StarScape](https://github.com/StarScape)
+  ([#1656](https://github.com/birchill/10ten-ja-reader/issues/1656)).
 - Made copying to the clipboard include rare headwords again when not using
   simplified copy mode
   ([#1665](https://github.com/birchill/10ten-ja-reader/issues/1665)).